sid/requirements.txt
apscheduler==3.10.4
beautifulsoup4==4.12.3
chromadb==0.5.5
faster-whisper==1.1.0
huggingface-hub==0.26.2
httpx==0.27.2
imagehash==4.3.1
mss==10.0.0
numpy==1.26.4
onnxruntime==1.20.1
openwakeword==0.6.0
Pillow==10.4.0
pydantic==2.10.3
PyAudio==0.2.14
psutil==6.1.1
pyautogui==0.9.54
pygetwindow==0.0.9
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-WebEngine==6.7.0
pystray==0.19.5
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.24.0
readability-lxml==0.8.1
sentence-transformers==3.3.1
silero-vad==5.1.2
sounddevice==0.5.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
watchdog==6.0.0
yt-dlp==2025.1.15
# Optional: Training dependencies (install manually)
# unsloth==2024.1
# trl==0.12.0
# peft==0.13.0
# accelerate==1.1.0
